refactor: architectural refactor
