+++ src/schema/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the schema directory a Python package.
